fix alloc

+++ test/count.c
@@ -1,7 +1,7 @@
 int putchar(int c);
 
 int main() {
-    float x = 0.0;
+    int x = 0;
     while (x < 10.0) {
         int y = x;
         putchar(y % 10 + '0');
